Move files to make diff more readable
