+++ src/app/clusters/operational-state-server/CodegenIntegration.cpp
@@ -19,7 +19,7 @@
  * @file
  * @brief Implementation for the Operational State Server Cluster
  ***************************************************************************/
-#include "OperationalStateCluster.h"
+#include "CodegenIntegration.h"
 
 #include <app-common/zap-generated/attributes/Accessors.h>
 #include <app-common/zap-generated/callback.h>
