+++ src/app/clusters/operational-state-server/app_config_dependent_sources.cmake
@@ -16,8 +16,8 @@
 TARGET_SOURCES(
   ${APP_TARGET}
   PRIVATE
-    "${CLUSTER_DIR}/OperationalStateCluster.cpp"
-    "${CLUSTER_DIR}/OperationalStateCluster.h"
+    "${CLUSTER_DIR}/CodegenIntegration.cpp"
+    "${CLUSTER_DIR}/CodegenIntegration.h"
     "${CLUSTER_DIR}/OperationalStateDelegate.h"
     "${CLUSTER_DIR}/operational-state-cluster-objects.h"
     "${CLUSTER_DIR}/operational-state-server.h"
