+++ src/app/clusters/operational-state-server/app_config_dependent_sources.gni
@@ -12,8 +12,8 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 app_config_dependent_sources = [
-  "OperationalStateCluster.cpp",
-  "OperationalStateCluster.h",
+  "CodegenIntegration.cpp",
+  "CodegenIntegration.h",
   "OperationalStateDelegate.h",
   "operational-state-cluster-objects.h",
   "operational-state-server.h",
